refactor(cherry-assistant): narrow feedback boundaries
