+++ src/main/ai/agents/builtin/builtinAgentDefinition.ts
