+++ src/main/ai/agents/ensureBuiltinAssistant.ts
@@ -3,7 +3,7 @@ import { agentService } from '@data/services/AgentService'
 import type { AgentEntity } from '@shared/data/api/schemas/agents'
 import type { UniqueModelId } from '@shared/data/types/model'
 
-import { loadBuiltinAssistantDefaults } from './builtin/BuiltinAgentDefinition'
+import { loadBuiltinAssistantDefaults } from './builtin/builtinAgentDefinition'
 
 export function ensureBuiltinAssistant(): AgentEntity {
   const defaults = loadBuiltinAssistantDefaults()
