--- src/renderer/pages/agents/feedback/index.ts
@@ -1,8 +1,0 @@
-export type { FeedbackComposerLaunch } from './composerLaunch'
-export {
-  clearFeedbackComposerLaunch,
-  FEEDBACK_INTENT_GUARD_TTL_MS,
-  getFeedbackIntentGuardCacheKey,
-  persistFeedbackComposerLaunch,
-  readFeedbackComposerLaunch
-} from './composerLaunch'
