+++ src/renderer/pages/settings/FeedbackDialog.tsx
@@ -89,8 +89,8 @@ export function FeedbackDialog({ open, onOpenChange }: FeedbackDialogProps) {
 
   const openAgentFeedback = async () => {
     try {
-      const session = await ipcApi.request('ai.agent.feedback_session.create')
-      openRoute(getFeedbackAgentRoute(session.id))
+      const { sessionId } = await ipcApi.request('ai.agent.feedback_session.create')
+      openRoute(getFeedbackAgentRoute(sessionId))
     } catch (error) {
       logger.error('Failed to create Cherry Assistant feedback session', error as Error)
       toast.error(t('settings.about.feedback.agent_error'))
