+++ src/shared/ipc/schemas/ai.ts
@@ -16,7 +16,6 @@ import {
   ScheduledTaskEntitySchema,
   TimeoutMinutesAtomSchema
 } from '@shared/data/api/schemas/agents'
-import { AgentSessionEntitySchema } from '@shared/data/api/schemas/agentSessions'
 import { AgentSessionWorkspaceSourceSchema } from '@shared/data/api/schemas/agentWorkspaces'
 import { JobScheduleNameAtomSchema, TriggerSchema } from '@shared/data/api/schemas/jobs'
 import { CleanupPolicySchema, type FileEntry, FileEntrySchema } from '@shared/data/types/file'
@@ -251,7 +250,7 @@ export const aiRequestSchemas = {
   }),
   'ai.agent.feedback_session.create': defineRoute({
     input: z.void(),
-    output: AgentSessionEntitySchema
+    output: z.strictObject({ sessionId: z.string().min(1) })
   }),
   'ai.agent.session.prewarm': defineRoute({
     input: z.strictObject({ sessionId: z.string().min(1) }),
